+++ src/handlers/subscribe.rs
@@ -4,11 +4,13 @@ use axum::{
     http::{HeaderMap, StatusCode},
     Json,
 };
+use axum_extra::extract::CookieJar;
 use std::sync::Arc;
 
 pub(crate) async fn subscribe(
     State(context): State<Arc<LongPollingServiceContext>>,
     headers: HeaderMap,
+    jar: CookieJar,
     Json([message]): Json<[Message; 1]>,
 ) -> HandlerResult<Json<[Message; 1]>> {
     tracing::info!("Got subscribe request: `{message:?}`.");
@@ -24,14 +26,19 @@ pub(crate) async fn subscribe(
     if channel.as_deref() != Some("/meta/subscribe") {
         Err(Message::session_unknown(id, channel, None).into())
     } else {
+        let session_unknown = || Message::session_unknown(id.clone(), channel.clone(), None);
+
+        let client_id = client_id.ok_or_else(session_unknown)?;
+        context
+            .check_client(&jar, &client_id)
+            .await
+            .ok_or_else(session_unknown)?;
+
         let subscription = subscription.ok_or_else(|| Message::subscription_missing(id.clone()))?;
 
         if subscription.is_empty() {
             return Err(Message::subscription_missing(id).into());
         }
-
-        let client_id =
-            client_id.ok_or_else(|| Message::session_unknown(id.clone(), channel.clone(), None))?;
 
         subscription.iter().try_for_each(|name| {
             context
@@ -43,8 +50,7 @@ pub(crate) async fn subscribe(
 
         context
             .subscribe(client_id, headers, subscription.clone())
-            .await
-            .map_err(|_| Message::session_unknown(id.clone(), channel.clone(), None))?;
+            .await;
 
         Ok(Json([Message {
             subscription: Some(subscription),