+++ src/types.rs
@@ -2,6 +2,8 @@ mod callback;
 mod client_id;
 mod client_receiver;
 mod client_sender;
+mod cookie_id;
+mod id;
 mod subscription_id;
 
 mod callback_args;
@@ -10,5 +12,7 @@ pub mod error;
 /// Contains cometd Message struct.
 pub mod messages;
 
-pub(crate) use {callback::*, client_receiver::*, client_sender::*, subscription_id::*};
+pub(crate) use {
+    callback::*, client_receiver::*, client_sender::*, cookie_id::*, id::*, subscription_id::*,
+};
 pub use {callback_args::*, client_id::*};